Queue: Use context menu for token card options and refactor code
