+++ public/locale/en.json
@@ -3121,6 +3121,7 @@
   "no_schedule_templates_found": "No schedule templates found for this month.",
   "no_scheduled_exceptions_found": "No scheduled exceptions found",
   "no_sections": "No sections added. Click \"Add Section\" to begin configuring your report sections.",
+  "no_service_points_available": "No Service Points Available",
   "no_service_points_found": "No service points found",
   "no_service_points_found_description": "No service points have been created yet. Create your first service point to get started.",
   "no_service_requests_found": "No service requests found",
@@ -4923,6 +4924,7 @@
   "todays_queues": "Today's Queues",
   "toggle_filters": "Toggle Filters",
   "token": "Token",
+  "token_assigned_to_service_point": "Token has been assigned to service point",
   "token_booking": "Token Booking",
   "token_categories": "Token Categories",
   "token_category": "Token Category",
